% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{DENSITY}
\alias{DENSITY}
\title{Function density}
\usage{
DENSITY(S, T_C)
}
\arguments{
\item{S}{Salinity in g/kg or pss (numeric)}

\item{T_C}{Temperature in degrees C (numeric)}
}
\value{
Density in g/cc
}
\description{
Equation of State is from Millero & Poisson (1981) DSR V28: 625-629.

  INPUT: Salinity (S) in g/kg or pss.
         Temperature (T) in degrees C.

  OUTPUT: Density [rho] in g/cc.

  DEFINE CONSTANTS FOR EQUATION OF STATE
}
