% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepstartmult.R
\name{grep.start.mult}
\alias{grep.start.mult}
\title{Function grep.start.mult}
\usage{
grep.start.mult(sub_str, str_list)
}
\arguments{
\item{sub_str}{vector of strings}

\item{str_list}{list of strings}
}
\value{
If found, an index value. Otherwise, NA
}
\description{
Returns the index (from str_list) of the firt matching string in sub_str
}
\examples{
l = c('hi','bye','sigh','rye','nigh')
sub = c('I','know','what','this','bye','function','sigh','does')
grep.start.mult(sub,l)
}
