% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcorrection.R
\name{ph.omega}
\alias{ph.omega}
\title{Compute ph correction based on CTD and Bottle data}
\usage{
ph.omega(
  ctd,
  bottle,
  flag = 8,
  k1k2 = "l",
  kf = "dg",
  ks = "d",
  pHscale = "T",
  b = "l10"
)
}
\arguments{
\item{ctd}{a data frame with AT LEAST the following information with these headings:

   \code{season} - the season;

   \code{agency} - the sampling agency;

   \code{sampledate} - the date of sample collection;

   \code{sampletime} - the time of sample collection;

   \code{station} - a code identifying the station;

   \code{depth} - An integer representing depth in feet;

   \code{fieldrep} - Field replicate;

   \code{labrep} - lab replicate;

   \code{Latitude} - latitude in decimal degrees.;

   \code{Longitude} - longitude in decimal degrees. Make sure there is a negative sign for the Western coordinates;

   \code{temperature} - Temperature in Celsius

   \code{salinity} - the salinity observed at the location in PSU, ideally at time of sampling.

   \code{density} - density;

   \code{ph} - ph;}

\item{bottle}{\code{season} - the season;

   \code{agency} - the sampling agency;

   \code{sampledate} - the date of sample collection;

   \code{sampletime} - the time of sample collection;

   \code{station} - a code identifying the station;

   \code{depth} - An integer representing depth in feet;

   \code{fieldrep} - Field replicate;

   \code{labrep} - lab replicate;

   \code{bottle} - a number identifying the bottle

   \code{lims} - not sure;

   \code{cruise} - not sure;

   \code{temperature} - Temperature in Celsius

   \code{salinity} - the salinity observed at the location in PSU, ideally at time of sampling.

   \code{ta} - alkalinity;

   \code{ph} - ph;}
}
\description{
Given CTD and Bottle data, merge the two dataframes together to calculate delta pH and get the corrected pH
}
\details{
Get the actual delta pH from the pHinsi function from seacarb
  Create a linear model to approximate delta ph for the other depths that dont have measurements
}
