% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH_nbs2total.R
\name{pH_nbs2total}
\alias{pH_nbs2total}
\title{Function pH_nbs2total}
\usage{
pH_nbs2total(S, T_C, pH_NBS, rho_SW)
}
\arguments{
\item{S}{Salinity in g/kg or pss (numeric)}

\item{T_C}{Temperature in degrees C (numeric)}

\item{pH_NBS}{pH measured on NBS scale}

\item{rho_SW}{density of SW at S, T_C in g/cc}
}
\value{
pH on Total Scale
}
\description{
CONVERT pH (ON NBS SCALE) TO pH on TOTAL SCALE
source: https://www.mbari.org/products/research-software/visual-basic-for-excel-oceanographic-calculations/
  link: https://www.mbari.org/wp-content/uploads/2016/03/pH_nbs2total.txt

This code was adapted from the work of others:
    CO2SYS.exe was originally written by Ernie Lewis and Doug Wallace.
    It was converted to Matlab by Richard E. Zeebe & Dieter A. Wolf-Gladrow.
    Parts were extracted and modified by Rachel M. Dunk.

I am indebted to all of them for their initial development of this code.

FIRST: convert pH_NBS to pH_SWS
  aH = 10^(-pH_NBS) = fH*H_SWS
  fH = activity coefficient of H+ ion & includes liquid junction effects
  The fit used in CO2SYS is valid for salinity 20-40 (from Takahashi et al,
    Chapter 3 in GEOSECS Pacific Expedition,v.3, 1982, p. 80).
  For future ref, CO2SYS assumes fH is independent of P
}
